% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_pattern.R
\name{count_rows_with_pattern}
\alias{count_rows_with_pattern}
\title{Count Rows Containing a Specific Pattern Using grep -c}
\usage{
count_rows_with_pattern(pattern, data = NULL, ignore_case = FALSE)
}
\arguments{
\item{pattern}{A character string specifying the pattern to search for within
the dataset rows. Passed directly to \code{grep}.}

\item{data}{A data frame or data.table in which to search for the pattern.
If \code{NULL} (the default), it will try to use \code{ggplot2::diamonds}.}

\item{ignore_case}{Logical. If \code{TRUE}, the \code{-i} flag is added to the \code{grep}
command for case-insensitive matching (default: \code{FALSE}).}
}
\value{
An integer representing the number of rows in the data where the
pattern is found. Returns \code{0L} if the pattern is not found or no matches are
returned.
}
\description{
This function writes the input data frame to a temporary file,
uses the system's \code{grep -o} command to count the occurrences of the
user-specified pattern, and returns the count. It's designed to be
memory-efficient for large datasets compared to reading matches into R.
}
